requirements.txt
aiohttp>=3.8.5
tenacity>=8.2.2
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dateutil>=2.8.2
pydantic>=1.10.9
pytest>=7.4.0
flake8>=6.1.0
requests>=2.31.0
tqdm>=4.66.1
# optional for demo (do not required for scraper)
sentence-transformers>=2.2.2
qdrant-client>=1.6.0

streamlit>=1.22.0
scikit-learn>=1.2.0
openai>=0.27.0

